+++ middlewares/validation/index.js
@@ -1,0 +1,5 @@
+const validation = require('./joi')
+
+module.exports = {
+  validation,
+}
