+++ model/contacts/listContacts.js
@@ -1,4 +1,4 @@
-const contacts = require('../contacts.json')
+const contacts = require('../../db/contacts.json')
 
 const listContacts = async () => {
   return contacts
