+++ model/contacts/removeContact.js
@@ -8,7 +8,7 @@ const removeContact = async contactId => {
   if (idx === -1) return null
 
   contacts.splice(idx, 1)
-  await updateContacts(contacts)
+  updateContacts(contacts)
 
   return true
 }
