+++ model/contacts/updateContacts.js
@@ -1,7 +1,7 @@
 const fs = require('fs/promises')
 const path = require('path')
 
-const contactsPath = path.join(__dirname, '..', 'contacts.json')
+const contactsPath = path.join(__dirname, '../../db/contacts.json')
 
 const updateContacts = async contacts => {
   await fs.writeFile(contactsPath, JSON.stringify(contacts))
